Extract share modal and fix button
