+++ app/javascript/app/components/modal-share/modal-share-actions.js
@@ -1,0 +1,7 @@
+import { createAction } from 'redux-actions';
+
+const setShareModal = createAction('setShareModal');
+
+export default {
+  setShareModal
+};
