+++ app/javascript/app/components/modal-share/modal-share-reducers.js
@@ -1,0 +1,12 @@
+export const initialState = {
+  isOpen: false
+};
+
+const setShareModal = (state, { payload }) => ({
+  ...state,
+  isOpen: payload.open
+});
+
+export default {
+  setShareModal
+};
